backend/requirements.txt
fastapi
uvicorn
openai
sqlalchemy
psycopg2-binary
python-dotenv